Extract logic out of Inspector
