+++ packages/replay-next/components/inspector/Inspector.module.css
@@ -4,10 +4,3 @@
   white-space: nowrap;
   overflow: hidden;
 }
-
-.InspectorWrapper {
-  padding: 0.25rem;
-  display: flex;
-  flex-direction: column;
-  overflow-x: hidden;
-}
