+++ src/ui/components/SecondaryToolbox/ReduxDevTools.module.css
@@ -143,3 +143,10 @@
   width: 100%;
   cursor: pointer;
 }
+
+.RDTInspector {
+  padding: 0.25rem;
+  display: flex;
+  flex-direction: column;
+  overflow-x: hidden;
+}
